# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/documents": {
      "post": {
        "operationId": "DocumentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Document 创建成功并入队",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "400": {
            "description": "校验失败"
          },
          "401": {
            "description": "未认证"
          },
          "404": {
            "description": "KB 不存在或无权访问"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建 Document",
        "tags": [
          "Documents"
        ]
      }
    },
    "/documents/{id}": {
      "get": {
        "operationId": "DocumentsController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Document ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document 详情",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证"
          },
          "404": {
            "description": "Document 不存在或无权访问"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取 Document 详情",
        "tags": [
          "Documents"
        ]
      }
    },
    "/kbs/{kbId}/documents": {
      "get": {
        "operationId": "DocumentsController_listByKb",
        "parameters": [
          {
            "name": "kbId",
            "required": true,
            "in": "path",
            "description": "KB ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document 列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentListResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证"
          },
          "404": {
            "description": "KB 不存在或无权访问"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "列出 KB 下 Document",
        "tags": [
          "Documents"
        ]
      }
    }
  },
  "info": {
    "title": "DevBrain API",
    "description": "DevBrain 知识库 API",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "devbrain_refresh"
      }
    },
    "schemas": {
...
      "CreateDocumentDto": {
        "type": "object",
        "properties": {
          "kbId": {
            "type": "string",
            "description": "目标 KB ID"
          },
          "objectKey": {
            "type": "string",
            "description": "对象存储 key"
          },
          "uploadToken": {
            "type": "string",
            "description": "上传签名 token"
          },
          "originalName": {
            "type": "string",
            "description": "原始文件名"
          },
          "mimeType": {
            "type": "string",
            "description": "MIME 类型"
          },
          "sizeBytes": {
            "type": "number",
            "description": "文件大小（字节）"
          }
        },
        "required": [
          "kbId",
          "objectKey",
          "uploadToken",
          "originalName"
        ]
      },
      "DocumentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Document ID"
          },
          "kbId": {
            "type": "string",
            "description": "所属 KB ID"
          },
          "sourceType": {
            "type": "string",
            "description": "来源类型"
          },
          "originalName": {
            "type": "string",
            "description": "原始文件名"
          },
          "status": {
            "type": "string",
            "description": "处理状态"
          },
          "errorCode": {
            "type": "object",
            "description": "错误码"
          },
          "errorMessage": {
            "type": "object",
            "description": "错误信息"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "创建时间"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "更新时间"
          }
        },
        "required": [
          "id",
          "kbId",
          "sourceType",
          "originalName",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "DocumentListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "description": "Document 列表",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentResponse"
            }
          }
        },
        "required": [
          "items"
        ]
      }
    }
  }
}